Extract ExportPreset resource into trait
Add ExportPresetResourceDefinition trait that centralizes model, navigation, form and table definitions for ExportPreset. Replace duplicated implementations in ExportPresetResourceV3 and ExportPresetResourceV4 to use the new trait. Update tests/TestCase.php to apply the real migration (2026_05_26_create_export_presets_table.php) instead of creating the export_presets table inline, keeping schema setup consistent with migrations.
